Refactor classes
Create controllers:
-books_controller - and factor out methods from app.rb to books_controller
-people_controller - and factor out methods concerned with people class to people_controller
-rentals_controller - test
-debug linters
