+++ rentals_controller.rb
@@ -1,0 +1,1 @@
+test
